requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.21.31
botocore==1.24.31
Django==4.0.3
django-appconf==1.0.5
django-extra-field-validation==1.1.1
django-imagekit==4.1.0
django-modeltranslation==0.17.5
django-nested-admin==3.4.0
django-quill-editor==0.1.40
django-solo==2.0.0
django-storages==1.12.3
djangorestframework==3.13.1
future==0.18.2
jmespath==1.0.0
pilkit==2.0
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-monkey-business==1.0.0
pytz==2022.1
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
